app/src/main/kotlin/com/dg/eqs/core/definition/relation/alteration/shiftingstep/matcher/DenominatorOutOfDivisionToPositiveProductMatcher.kt
package com.dg.eqs.core.definition.relation.alteration.shiftingstep.matcher

import com.dg.eqs.core.definition.relation.alteration.shiftingstep.ShiftingMatcher
import com.dg.eqs.core.definition.term.Term
import com.dg.eqs.core.definition.term.Terms
import com.dg.eqs.core.definition.term.operation.dotoperation.division.Division
import com.dg.eqs.core.definition.term.operation.dotoperation.product.PositiveProduct


object DenominatorOutOfDivisionToPositiveProductMatcher : ShiftingMatcher {
    override fun matches(sourceSide: Term, targetSide: Term, source: Terms): Boolean {
        if(sourceSide !is Division) return false
        if(targetSide !is PositiveProduct) return false

        return sourceSide.denominator in source
    }
}